# backend/routes/api.yaml
swagger: "2.0"
info:
  title: BrainBeatsWeb API 
  description: This is a BrainBeatsWeb API documentation form
  version: 1.0
  licenses:
    name: Apache 2.0
    url: "http://www.apache.org/liscences/LIS"
  servers:
    -url: http://api.brainbeatz.xyz #https://api.brainbeats.dev
    description: Local server
schemes: [http] #[https]
...
paths:
...
    /api/users/createUser:
      post:
        tags:
          - Users
        summary: Create a user
        operationId: createUser
        description: Create a new user in the database
        consumes: 
        - application/json
        produces:
        - application/json
        parameters:
        - in: body
          name: Create
          required: true
          description: User Creation
          schema:
            $ref: '#/definitions/createUser'
        responses:
          200:
            description: OK
          404:
            description: URL Not Found

  # Find user by userID
    /api/users/getUserByID/:
      get:
        tags:
          - Users
        summary: Search a user by with their username
        operationId: getUserByID
        description: Search a user in the database
        consumes: 
        - application/json
        produces:
        - application/json
        parameters:
        - in: query
          name: id
          schema:
            type: string
          description: Enter a user UUID
          schema:
            type: string
        responses:
          200:
            description: OK
          404:
            description: URL Not Found

  # Find user by username
    /api/users/getUserByUsername/:
      get:
        tags:
          - Users
        summary: Search a user by with their username
        operationId: getUserByUsername
        description: Search a user in the database
        consumes: 
        - application/json
        produces:
        - application/json
        parameters:
        - in: query
          name: username
          schema:
            type: string
          description: Enter a user username
          schema:
            type: string
        responses:
          200:
            description: OK
          404:
            description: URL Not Found

      # Find user by reset password token
...
    /api/users/updateUser:
      put:
        tags:
          - Users
        summary: Update user
        operationId: updateUser
        description: Update user information
        consumes: 
        - application/json
        produces:
        - application/json
        parameters:
        - in: body
          name: Update
          required: true
          description: UpdateUser
          schema:
            $ref: '#/definitions/updateUser'
        responses:
          200:
            description: OK
          404:
            description: URL Not Found

  # Delete a user by ID
    /api/users/deleteUser:
      delete:
        tags:
          - Users
        summary: Delete user by username
        operationId: deleteUser
        description: Delete a user from the database
        consumes:
        - application/json
        produces:
        - application/json
        parameters:
        - in: body
          name: Delete User
          required: true
          description: Delete
          schema:
            $ref: '#/definitions/deleteUser'
        responses:
          200:
            description: OK
          404:
            description: URL Not Found



  # ***** User Tracks ***** #

  # Returns an array of all user tracks
...
definitions:
...
    createUser:
      required: 
      - firstName
      - lastName
      # - dob
      - email
      - username
      - password
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        # dob: 
        #   type: date 
        #   example: 1995-03-25
        email:
          type: string
          example: JohnDoe@gmail.com
        username:
          type: string
          example: JohnDoe96
        password:
          type: string
          example: JohnDoePassword
    updateUser:
      required: 
      - id
      - firstName
      - lastName
      - email
      - username
      properties:
        id:
          type: string
          example: 632961af-29ae-4c3a-8f3f-426763a4cf00
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        # dob: 
        #   type: date
        #   exmaple: 1997-07-13T00:00:00.000Z
        email:
          type: string
          example: JohnDoe@gmail.com
        username:
          type: string
          example: JohnDoe96
    deleteUser:
      required:
      - id
      properties:
        id:
          type: string
          exmaple: 632961af-29ae-4c3a-8f3f-426763a4cf00

  # ***** User Tracks ***** #
...